refactor : useSearchResultInfo 훅
-
